tests/requirements.txt
pytest==6.2.2
pytest-cov==2.11.1
PyYAML==5.4.1